chore: move base dht to dht __init__
